+++ MaxPowerLevel/Services/IRecommendations.cs
@@ -5,7 +5,7 @@ namespace MaxPowerLevel.Services
 {
     public interface IRecommendations
     {
-        IEnumerable<string> GetRecommendations(IEnumerable<Item> items, int powerLevel);
-        IEnumerable<Engram> GetEngramPowerLevels(int powerLevel);
+        IEnumerable<string> GetRecommendations(IEnumerable<Item> items, decimal powerLevel);
+        IEnumerable<Engram> GetEngramPowerLevels(decimal powerLevel);
     }
 }
